HanselGretel dialogues inserted

+++ compsoc-ascii-gamejam/Story/HanselGretel.cs
@@ -12,51 +12,80 @@ public class HanselGretel //start the story for hansel and gretel
     private HanselGretel()
     {
         /*
-        From a window in the cottage character will see hansel and gretel getting up
-        Confused character continues to walk to the cottage in order to make sure granny is alright and ask why there are children inside
-        Character walks in to find Hansel about to shove granny into the oven
 
-        RUN:
-        Character runs away to get help
-        Granny dies, character returns with help but the children are gone
+        As you approach your granny's house you see 2 children in the upstairs window. Confused you speed up your pace slightly in order to make sure your granny is okay as you've
+        never seen these children before. You walk in to see your granny with her head in the oven and the little boy about to shove her in. Do you "run" off to go get help, begin 
+        trying to "talk" to the children and find out what they're doing, or do you immediately go to "fight" in order to protect your poor frail old granny?
 
-        TALK:
-        Kids explain what happened
-        Character gets choice on wether or not to believe them
-            Believe 
-            Character is mortified and lets kids continue in shock finishing the story and go into an outro of character being in shock about what happened and eventually goes home
-            crying to their mother
+        *GET USER INPUT*
 
-            Don't Believe
-            Character will say to the kids there must have been a misunderstanding and tells them to leave granny alone
-            Charisma roll if char. wins kids will leave
-            Character will comfort granny and give her the foood and granny acts very thankful and says that she has no idea who those kids were acting all innocent
-            CHaracter goes home and good ending occurs
+        IF RUN
+            You run away to go and get help for you granny. By the time you return she is gone, and the children are nowhere to be seen. You stare in shock unable to believe what
+            has just occured and question if you made the right choice. 
+
+            *GAME END, THIS IS THE TRUE ENDING TO HANSEL AND GRETEL THAT YOU HAVE JUST DISCOVERED*
+
+        IF TALK
+        
+        The kids begin to explain to you that your granny was going to eat them. You stare at them, and look between them and your sweet frail old granny unsure who to believe.
+        Do you believe the kids?
+
+        IF YES
+
+            You stand there mortified as the kids finish their story, about what your granny had done. This causes you to run away in shock and go home explaining what just happened
+            to your mother who while suspicious doesn't press you or lecture you about lying.
+
+        IF NO
+        
+            You laugh at the kids and tell them that's ridiculous and that they should just leave her house and stop harassing the poor old lady. The kids stare at you before realizing
+            that they have an opportunity to escape and run for freedom. You help granny to her living room where she is distraught insisting she has no idea who those kids were, or how
+            they got into her house. You sigh, and comfort her until she falls asleep where you then return home, and go to bed yourself.
 
         FIGHT:
         Character immediately attacks kids, combat mechanic occurs, kids attack as one entity however each with low stats but added are quite strong
         Character will kill the children, however in the process drop the basket causing the fruit to fall on the floor and get stepped on so granny now has no food to eat
         in return granny will turn to the character
 
-         GRANNY IS NOW THE VILLAIN
-         RUN:
-         agility roll for character against granny
-         Character will run away from granny as fast as possible
-            Sucess:
-            Character successfully gets home and away from granny explains what happened and either transfer into other story or end with mum telling character not to make things up
-            and character gets sent to room ending game
-            Fail:
-            Granny catches up to character and then character loses 3 hp before escaping and continuing trying to make it home however granny will still be chasing character leading
-            to mum seeing and stepping in to save character
+        *GRANNY IS NOW THE VILLAIN*
+
+        With no food to eat now that her dinner is dead, and you dropped the gifts from your mother granny turns to you. You begin to back up slightly realizing that she intends to 
+        eat you now. That maybe you shouldn't have killed those kids and they were defending themselves. Her eyes glow red as you tremble in fear. Do you "Run", try to "talk" your way
+        out of this, or "fight" your own granny?
+
+        IF RUN:
+            
+            You begin to start running home as fast as you can. Nothing on your mind but surviving.
+
+            *INSERT AGILITY ROLL*
+
+            IF WIN
+                
+                You run until you've made it home, and begin to try explaining what happened to your mother. She rolls her eyes however and simply tells you to stop lying. When you 
+                insist that you're being truthful she sends you to your room and says that you're no longer allowed to explore as it's given you crazy ideas.
+
+                *GAME END*
+            
+            IF LOSS
+
+                You run however you trip giving granny a chance to catch up so it looks as though you two made it home together. However your mum sees you running and the terrifying
+                look in her eyes and steps in, throwing a knife at her, causing granny to scream out before she crumpled to the ground. Therefore saving you.
 
         TALK:
-        charisma roll against granny
-        Character will attempt to convince granny to go home and this will all remain a secret and character will explain to mother that they tripped on the way and lost the goodies
-            Success:
-            Granny agrees and they walk home however this means an evil is still remaining so that is a loss of a WP
-            Fail:
-            Granny says no and kills character resulting in a game over
+        
+            You begin to try and reason with granny. You beg for her to just let you go home, and that this can be your littel secret, you won't tell a soul. As far as your mother knows
+            you tripped on your way there and lost the basket of goods. 
 
+            *INSERT CHARISMA ROLL*
+
+            IF WIN
+
+                Granny luckily agrees with you, and allows you to go home.
+            
+            IF LOSS
+
+                Ganny laughs and informs you that she's hungry. Before you know it she has grabbed you and you feel a burning sensation hit you before everything goes dark.
+
+                *GAME OVER, YOU DEADED*
         FIGHT:
         fight mechanic
         Character will launch into self defense mode
